It is done.
The Events Tab now is pretty, logging everything.
Some statiscts were deemed unnecessary, so we got a few blank spaces,
but that should be ok.

+++ src/meta/BaseSimulator.java
@@ -178,5 +178,5 @@ public abstract class BaseSimulator
 	 * 
 	 * @return A integer value indicating the total time queues were empty.
 	 */
-	public abstract double getEmptyQueueTime();
+	public abstract int getEmptyQueueTime();
 }
